refactor: :rainbow: signup/local API 디렉토리 변경
signup이 auth 폴더에 존재하는 것이 부적절하다고 판단해서
user 이라는 라우터로 옮기고, 이에 따른 폴더 구조를 변경했습니다
